修正連結記錄頁面bug問題
