+++ index.html
@@ -17,7 +17,7 @@
     <title>蝦英雄</title>
   </head>
   <body>
-    <div id="app" class=" bg-surface text-main font-nunito"></div>
+    <div id="app" class="bg-surface/50 text-main font-nunito"></div>
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
