+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "shopee-short-link",
-  "version": "3.0.1",
+  "version": "3.1.1",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
@@ -20,17 +20,18 @@
     "@formkit/auto-animate": "^1.0.0-beta.3",
     "@vueuse/core": "^9.3.1",
     "@vueuse/integrations": "^9.13.0",
-    "pinia": "^2.1.3",
+    "pinia": "^2.1.6",
     "qrcode": "^1.5.3",
     "vue": "^3.2.26",
-    "vue-router": "^4.2.2"
+    "vue-router": "^4.2.4"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^4.2.3",
     "autoprefixer": "^10.4.14",
-    "postcss": "^8.4.24",
-    "tailwindcss": "^3.3.2",
-    "vite": "^4.3.9",
+    "postcss": "^8.4.27",
+    "tailwindcss": "^3.3.3",
+    "unplugin-vue-router": "^0.6.4",
+    "vite": "^4.4.9",
     "vite-plugin-radar": "^0.6.0"
   }
 }
