+++ src/stores/link.js
@@ -1,31 +1,14 @@
 import { defineStore } from 'pinia'
 
-const key = 'shortLink'
+const KEY = 'shortLink'
 
-/**
- * 設定localstorage資料
- */
 const LS = {
-	/**
-	 * 讀取localstorage資料
-	 * @param {String} listName key名稱
-	 * @returns {Array} 若無資料回空陣列,若有則回obj
-	 */
 	load (listName) {
 		return JSON.parse(localStorage.getItem(listName) || '[]')
 	},
-	/**
-	 * 寫入localstorage資料
-	 * @param {String} listName key名稱
-	 * @param {Object} data 要寫入localstorage的資料
-	 */
 	save (listName, data) {
 		localStorage.setItem(listName, JSON.stringify(data))
 	},
-	/**
-	 * 刪除localstorage資料
-	 * @param {string} listName key名稱
-	 */
 	remove (listName) {
 		localStorage.removeItem(listName)
 	}
@@ -40,15 +23,15 @@ export const useLinkStore = defineStore({
   },
   actions: {
     initState () {
-      this.links = LS.load(key)
+      this.links = LS.load(KEY)
     },
     addLink (linkData) {
       this.links.unshift(linkData)
-      LS.save(key, this.links)
+      LS.save(KEY, this.links)
     },
     deleteLink (index) {
       this.links.splice(index, 1)
-      LS.save(key, this.links)
+      LS.save(KEY, this.links)
     }
   }
 })
